+++ lib/android_media.dart
@@ -3,4 +3,10 @@
 /// Provides classes that manage various media interfaces in audio and video.
 ///
 /// See: https://developer.android.com/reference/android/media/package-summary
+///
+/// {@canonicalFor face.Face}
+/// {@canonicalFor face_detector.FaceDetector}
 library android_media;
+
+export 'src/media/face.dart' show Face;
+export 'src/media/face_detector.dart' show FaceDetector;