+++ lib/src/media/face.dart
@@ -1,0 +1,7 @@
+/* This is free and unencumbered software released into the public domain. */
+
+/// A [Face] contains all the information identifying the location of a face in
+/// a bitmap.
+///
+/// See: https://developer.android.com/reference/android/media/FaceDetector.Face
+abstract class Face {}