+++ lib/src/media/face_detector.dart
@@ -1,0 +1,8 @@
+/* This is free and unencumbered software released into the public domain. */
+
+import '../graphics/bitmap.dart' show Bitmap;
+
+/// Identifies the faces of people in a [Bitmap] graphic object.
+///
+/// See: https://developer.android.com/reference/android/media/FaceDetector
+abstract class FaceDetector {}